requirements.txt
urllib3==1.26.14
asgiref==3.4.1
Django==3.2.8
powerline-status==2.7
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
gunicorn==22.0.0
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
cos-python-sdk-v5==1.9.30